+++ src/SFML/Graphics/RenderWindow.cpp
@@ -39,7 +39,7 @@ RenderWindow::RenderWindow()
 
 
 ////////////////////////////////////////////////////////////
-RenderWindow::RenderWindow(VideoMode mode, const std::string& title, unsigned long style, const ContextSettings& settings)
+RenderWindow::RenderWindow(VideoMode mode, const std::string& title, Uint32 style, const ContextSettings& settings)
 {
     // Don't call the base class constructor because it contains virtual function calls
     Create(mode, title, style, settings);
